Moving implementations out of lbann.proto module.
